requirements.txt
acme==3.3.0
allure-pytest==2.15.0
allure-python-commons==2.15.0
amqp==5.1.1
annotated-types==0.7.0
APScheduler==3.6.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==25.3.0
billiard==4.1.0
cachetools==4.2.2
celery==5.3.1
certbot==2.11.1
certbot-dns-cloudflare==2.11.1
certifi==2025.8.3
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudflare==2.19.4
colorama==0.4.6
ConfigArgParse==1.7.1
configobj==5.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.6
defusedxml==0.7.1
distro==1.9.0
Django==4.2.13
django-admin-notification==0.2
django-cors-headers==4.7.0
django-filter==23.2
django-ratelimit==4.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
dnspython==2.7.0
drf-yasg==1.21.4
email-validator==2.3.0
frappe-manager==0.18.0
generics==6.0.0
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
h11==0.16.0
idna==3.4
inflection==0.5.1
iniconfig==2.1.0
inquirerpy==0.3.4
itypes==1.2.0
Jinja2==3.1.6
josepy==1.15.0
jsonlines==4.0.0
kombu==5.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ngrok==1.4.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.1
parsedatetime==2.6
passlib==1.7.4
pfzy==0.3.4
pillow==10.2.0
pluggy==1.6.0
prompt-toolkit==3.0.39
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pyOpenSSL==25.1.0
pyRFC3339==2.1.0
PySocks==1.7.1
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-telegram-bot==13.15
python3-openid==3.2.0
pytz==2023.3
pywin32==311
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
selenium==4.35.0
sentry-sdk==1.40.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
stripe==6.1.0
tomlkit==0.12.5
tornado==6.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.5.0
vine==5.0.0
wcwidth==0.2.6
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
